pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emtoflow"
version = "0.1.0"
description = "EMTO input generation and optimization workflows for electronic-structure calculations."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
  {name = "Pamela Costa Carvalho"}
]
dependencies = [
  "numpy>=1.19.0",
  "scipy>=1.5.0",
  "pandas>=1.3.0",
  "matplotlib>=3.3.0",
  "pyyaml>=5.4.0",
  "pymatgen>=2022.0.0",
]

[project.urls]
Homepage = "https://github.com/pcostacarvalho/emtoflow"

[project.scripts]
emtoflow-opt = "emtoflow.cli.run_optimization:main"
emtoflow-generate-percentages = "emtoflow.cli.generate_percentages:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["emtoflow*"]

